+++ packages/react/components/dropdown/group/index.ts
@@ -1,0 +1,2 @@
+export { default as DropdownGroup } from './DropdownGroup'
+export * from './DropdownGroupProps'