+++ packages/react/components/dropdown/item/index.ts
@@ -1,0 +1,2 @@
+export { default as DropdownItem } from './DropdownItem'
+export * from './DropdownItemProps'