+++ packages/react/components/radio/tiles/tile/RadioTileProps.ts
@@ -1,18 +1,20 @@
+import { VariantProps } from '@/objects'
 import { TouchableOpacity } from 'react-native'
 import { IconName, IconNameValues } from '../../../../components/icon'
-import { RadioProps } from '../../../../components/radio/RadioProps'
-import { VariantProps } from '@/objects'
+import { RadioNativeProps, RadioProps } from '../../../../components/radio/RadioProps'
 
 /**
  * Columns Item Interface
  */
-export interface RadioTileProps extends RadioProps {
+interface RadioTilePropsCommon {
   horizontal?: boolean
   icon?: IconName | IconNameValues
   description?: string | React.ReactNode
   sticker?: string
   stickerVariant?: VariantProps['variant']
 }
+export interface RadioTileProps extends RadioProps, RadioTilePropsCommon {}
+export interface RadioTileNativeProps extends RadioNativeProps, RadioTilePropsCommon {}
 
 export type RadioTileRef = HTMLDivElement
 export type RadioTileNativeRef = TouchableOpacity