+++ packages/react/components/textarea/Textarea.native.tsx
@@ -6,12 +6,12 @@ import {
   InputKeyboardType,
   InputTextContentType,
 } from '@/components/input/InputEnum'
-import { grayscale, TypographyColor } from '@/objects'
+import { TypographyColor } from '@/objects'
 import { getColorStyle, TrilogyColor } from '@/objects/facets/Color'
 import { StatusState } from '@/objects/facets/Status'
-import React, { useEffect, useRef, useState } from 'react'
-import { Animated, StyleSheet, Text, TextInput, View } from 'react-native'
-import { Spacer, SpacerSize } from '../spacer'
+import React, { useState } from 'react'
+import { StyleSheet, Text, TextInput, View } from 'react-native'
+import { SpacerSize } from '../spacer'
 import { Text as TrilogyText } from '../text'
 import { TextLevels } from '../text/TextEnum'
 import { TextareaNativeProps, TextareaNativeRef } from './TextareaProps'
@@ -29,10 +29,9 @@ import { TextareaNativeProps, TextareaNativeRef } from './TextareaProps'
  * @param ref Pass a ref for textarea
  * @param maxLength {number} Textarea max length
  * @param rows {number} Textarea rows
- * @param iconName {IconName | IconNameValues} display Icon
- * @param statusIconName {IconName | IconNameValues} display status Icon
+ * @param iconNameLeft {IconName | IconNameValues} display Icon on the left
+ * @param iconNameRight {IconName | IconNameValues} display Icon on the right
  * @param testId {string} Test Id for Test Integration
- * @param dynamicPlaceholder {boolean}
  * @param status {InputStatus} Textarea with status - (SUCCESS|WARNING|ERROR|DEFAULT)
  * @param keyboardStyle {InputKeyboardAppearance} Custom appearance for keyboard
  * @param autoCapitalize {InputAutoCapitalize} Capitalize => NONE | SENTENCES | WORDS | CHARS
@@ -61,7 +60,6 @@ const Textarea = React.forwardRef<TextareaNativeRef, TextareaNativeProps>(
       keyboardType,
       status,
       maxLength,
-      dynamicPlaceholder = true,
       label,
       iconNameLeft,
       iconNameRight,
@@ -73,27 +71,8 @@ const Textarea = React.forwardRef<TextareaNativeRef, TextareaNativeProps>(
     ref,
   ): JSX.Element => {
     const [_value, setValue] = useState<string>(value || '')
-
     const [isFocus, setIsFocus] = useState<boolean>(false)
-
-    const [displayDynamicLabel, setDisplayDynamicLabel] = useState<boolean>(false)
     const textareaColor = isFocus ? getColorStyle(TrilogyColor.MAIN) : getColorStyle(TrilogyColor.NEUTRAL)
-
-    const animation = useRef(new Animated.Value(0)).current
-
-    useEffect(() => {
-      if (displayDynamicLabel) {
-        Animated.timing(animation, {
-          toValue: 1,
-          duration: 5000,
-          useNativeDriver: false,
-        }).start()
-      }
-    }, [displayDynamicLabel, animation])
-
-    useEffect(() => {
-      setValue(value || '')
-    }, [value])
 
     const styles = StyleSheet.create({
       textarea: {
@@ -107,13 +86,12 @@ const Textarea = React.forwardRef<TextareaNativeRef, TextareaNativeProps>(
           textareaColor,
         height: customHeight,
         justifyContent: 'flex-start',
-        paddingLeft: iconNameLeft ? 48 : 16,
-        paddingRight: maxLength ? 48 : 16,
-        paddingTop: dynamicPlaceholder && displayDynamicLabel ? 24 : 8,
+        paddingLeft: (iconNameLeft && isFocus && 47) || (iconNameLeft && 48) || 16,
+        paddingRight: (iconNameRight && 48) || 16,
+        paddingTop: isFocus ? 10 : 11,
         textAlignVertical: 'top',
         color: getColorStyle(TrilogyColor.MAIN),
         backgroundColor: disabled ? getColorStyle(TrilogyColor.DISABLED_FADE) : getColorStyle(TrilogyColor.BACKGROUND),
-        /*  width: '',*/
       },
       help: {
         fontSize: 12,
@@ -135,56 +113,32 @@ const Textarea = React.forwardRef<TextareaNativeRef, TextareaNativeProps>(
         backgroundColor: disabled ? getColorStyle(TrilogyColor.DISABLED) : 'white',
         padding: 3,
       },
-      dynamicLabel: {
-        position: 'absolute',
-        top: 2,
-        left: iconNameLeft ? 40 : 8,
-        fontSize: 12,
-        color: grayscale(getColorStyle(TrilogyColor.FONT)),
-        backgroundColor: 'transparent',
-        padding: 8,
-        paddingBottom: 4,
-      },
       leftIcon: {
         position: 'absolute',
-        top:
-          (dynamicPlaceholder && 11) ||
-          (!dynamicPlaceholder && label && sample && 67) ||
-          (!dynamicPlaceholder && label && !sample && 42) ||
-          45,
+        top: (label && sample && 67) || (label && !sample && 42) || 45,
         left: 16,
         zIndex: 10,
       },
       rightIcon: {
         position: 'absolute',
-        top:
-          (dynamicPlaceholder && 11) ||
-          (!dynamicPlaceholder && label && sample && 67) ||
-          (!dynamicPlaceholder && label && !sample && 42) ||
-          45,
+        top: (label && sample && 67) || (label && !sample && 42) || 45,
         right: 16,
         zIndex: 10,
       },
     })
 
     return (
-      <View>
-        {!dynamicPlaceholder && label && (
-          <>
-            <TrilogyText typo={TypographyColor.TEXT_DISABLED}>
-              {label} {label && required && <TrilogyText typo={TypographyColor.TEXT_ERROR}>*</TrilogyText>}
-            </TrilogyText>
-            <Spacer size={SpacerSize.THREE} />
-          </>
+      <View style={{ gap: SpacerSize.THREE }}>
+        {label && (
+          <TrilogyText typo={TypographyColor.TEXT_MAIN}>
+            {label} {label && required && <TrilogyText typo={TypographyColor.TEXT_ERROR}>*</TrilogyText>}
+          </TrilogyText>
         )}
 
-        {!dynamicPlaceholder && label && sample && (
-          <>
-            <TrilogyText level={TextLevels.THREE} typo={TypographyColor.TEXT_DISABLED}>
-              {sample}
-            </TrilogyText>
-            <Spacer size={SpacerSize.THREE} />
-          </>
+        {sample && (
+          <TrilogyText level={TextLevels.THREE} typo={TypographyColor.TEXT_MAIN}>
+            {sample}
+          </TrilogyText>
         )}
 
         {iconNameLeft && (
@@ -206,7 +160,6 @@ const Textarea = React.forwardRef<TextareaNativeRef, TextareaNativeProps>(
           textContentType={textContentType || InputTextContentType.NONE}
           keyboardType={keyboardType || InputKeyboardType.DEFAULT}
           onChangeText={(text) => {
-            setDisplayDynamicLabel(text.length > 0)
             setValue(text)
             if (onChange) {
               onChange({
@@ -219,10 +172,10 @@ const Textarea = React.forwardRef<TextareaNativeRef, TextareaNativeProps>(
           style={styles.textarea}
           onFocus={() => setIsFocus(true)}
           onBlur={() => setIsFocus(false)}
-          {...others}
           placeholderTextColor={
             disabled ? getColorStyle(TrilogyColor.DISABLED) : getColorStyle(TrilogyColor.FONT_PLACEHOLDER)
           }
+          {...others}
         />
 
         {iconNameRight && (
@@ -231,7 +184,6 @@ const Textarea = React.forwardRef<TextareaNativeRef, TextareaNativeProps>(
           </Text>
         )}
 
-        {displayDynamicLabel && dynamicPlaceholder && <Text style={styles.dynamicLabel}>{label}</Text>}
         {maxLength && (
           <Text style={styles.counter}>{_value ? `${_value?.length} / ${maxLength}` : `0 / ${maxLength}`}</Text>
         )}