+++ CHANGELOG.md
@@ -3,15 +3,23 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.14.0] - 2025-12-XX
+## [4.14.2] - 2026-01-15
+- Fix container queries
+- Fix dropdown menu (web)
+- Fix input and textarea props
+
+## [4.14.0] - 2026-01-12
 - Add datepicker native
 - Add calendar native
 - Add container query (FlexBox web)
 - Fix calendar order days
 - Fix position on range calendar (web)
 - Fix radio checkbox stretched (web)
+- Update skeleton (native)
+- Dropdown Menu (web)
 - Update and fix skeleton (native)
 - Fix flexbox gap (mobile container query)
+- Fix onChange prop (Radio)
 
 ## [4.13.0] - 2025-11-24
 - Fix Tab native fullwidth (ui)